backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0